Fixed unmatched variable bug.
If a rule was using an action in the identifier slot that was neither tested nor created elsewhere, Visual Soar was not recognizing it as a bug.

The Soar parser doesn’t check for this at parse time.  I think the code that checks for this is in Soar’s main C++.  So I had to add substantial code to review each parsed production, compile a list of variables tested/create and a list of variables used on RHS, and then compare the two for discrepancies.