Implented Bug #27: Paste as Link
Bug #27: [Easy] Add a menu option to to the datamap editor window for copying a link to the current node so you can paste the link (equiv to ctrl + shift + drag)

This ended up being a bit trickier than "Easy" as I had to figure out how
the Swing drag-and-drop architecture worked and also some learning curve
with how the datamap is structured.

As part of this fix, I also added Copy/Paste/Link to the context menu for
easier access.