Complete code reformat
Notably I switched all code to the end-of-line brace style.  It doesn't matter much to me what style it is but it's important to be consistent.  I picked this style because it's more popular.

I also enforced that alls blocks be in braces except one-line if-statements.

Finally I removed all the .java_ files. These were presumably old versions of code that is no longer needed.