Fixed bug #30:  auto-justify glitch
VS was getting confused by productions like this:

sp {foo*propose*test
   (state <s> ^name foo)
   {(<s> ^test abc)}
  -(<s> ^test2 abc)
-->
   (<s> ^operator <op> + =)
   (<op> ^name test)
}

sp {propose*initialize-example
   (state <s> ^superstate nil
             -^name)
-->
   (<s> ^operator <o> +)
   (<o> ^name initialize-example)
}

The root causes where:
1.  getElementString() was sometimes returning a string with a non-ASCII
    char in position 0.  I've no idea how that's happening but I suspect
    it's a bug in Java's AbstractDocument class which I can't fix so I
    had to detect and repair on our side.
2.  justifyDocument() was misisng a special cause for a negated condition
    after a compound condition.  No surprise that was overlooked.

Overall, justifyDocument() is a mess.  It's basically a hand-crafted
parser that likely started simple and then, each time a new bug is found,
a new condition has be added to handle it.  It really ought to be
revamped but doing so would not only take time but would likely
be much more buggy before it got better.  Alas, we'll likely never
be able to invest in that especially without unit tests.