New persistent preferences
The divider between the operator pane and the desktop is much closer to the left.  Furthermore, if you move it VS will “remember” where you set it and keep it there every time you run the app.

I’ve also fixed a minor bug in how projects are loaded.  There is support for the Open Project dialog to begin in the folder where you opened a project last but a key line of code was missing.

I also have VS remembering what font size setting