Fixed bug:  bad feedback location
When you perform a local datamap check for a single file via the
"Soar->Check All Productions" menu option in an operator window
AND there is a syntax error then the check aborts and the syntax
error is reported in the feedback window.

So far this is exactly what should happen.  The problem was that
the FeedbackListObject associated with the error did not have
an associated file or line number as it should.  So, when you
double click on it you get an erroneous result.