VS should not save the project if no change has been made
Before you simply had to load a project and close it again to cause
the main files (.vsa, main .soar and .dm files) to be written out
again.  This refreshes the timestamp on those files which causes
issues with backups.

Now there is a static variable in CustomInternalFrame that tracks
if _any_ file as _ever_ been changed since the project was opened.
No changes will be made to the main files at project exist unless
this flag has been tripped.

One other change sneaked in here:  I made yet more improvements
to the "remember where the divider was" code.  It's proven a lot
trickier than I thought!  There is one more issue with it to fix
and that will be in a future revision.