Fixed Highlighted Text Replace Bug
If you highlight text and then type a key syntactical character
(e.g., '<' or '.') then VS was correctly replacing the highlighted
text with the new char but then placing the cursor as if the
the highlighted text had not been deleted.

I added code to detect this situation and handle it properly.