Added 'Create Top-Level Folder' item to Operator Pane's Context Menu
I have the handler method setup but the "meat" of it is TBD.
I'm commiting now because in the process of doing this I discovered
I needed to refactor the code that enabled/disables context menu
items.  It was copy/pasted so badly I couldn't tell heads from tails
so I modified the operator node hierarchy to have it inherit
enable/disable settings so as to make the differencs clear and concise.

Hopefully no bugs introduced by this refactor.