Revamp Feedback list objects
The FeedbackListObject class was getting unwieldy and adding support
for ForeignVertex validation was going to make it worse.  So, I bit
the bullet and broke it up into parent + sub classes.

There's a little code in this revsision for the FV validation too
but it's not being called anywhere so it's just a stub atm.