Added Re-Import subtree option for linked datamaps
A Re-Import Subtree option is not in the context menu for level-1
datamap entries that are ForeignVertex (ref to external datamap).

Implementing this required revamping the addForeignSubTree() method
to not use FakeTreeNode objects.  I also moved it from
CheckBoxDataMapTree to the parent DataMapTree.

Some stub code exists (but is not active) for comparing an imported
subtree to its external original.  I still think it would be wise
to do this automatically whenever you a load a project that has
foreign vertexes just to keep things in sync.  For now, it's up
to the user to know it's out of date and re-import it.

This revision contains a few other minor code adjustments.