Fixed bugs in .cfg file loading
bug #1:  windows were being closed before the commit so the .cfg
file was being written as empty.  I worry that by commiting before
closing the windows I may have broken something.  I can unbreak it
later if need be by saving the .cfg file before closing all the
windows and not saving the .cfg file in the commit action.  It makes
ctrl+s/File-Save action not save the .cfg file this way but I could
likely find a way to work around that (e.g., a boolean flag indicating
if commit is user-initiated or the side effect of a Close).

bug #2:  When loading the .cfg file the same file could have different
filenames so it was not found.  To fix this I had the readCfgFile()
method convert all files to their canonical form.