VisualSoar autoloads new operators
When adding a new sub-operator file (via context menu on a node in the operator pane) VS now automatically opens the newly created file and auto-inserts the propose and apply operator templates.

This was a trickier change than it appears.  I had to change the signature of a couple methods in the OperatorNode class.  I think it will be fine but I'm slightly ware of potential issues that may arise.