Visual Soar warns about unmatched angle brackets
When an unclosed angle bracket is used as a constant in Soar it's
legal syntax, but it's likely not intentional.  For example:
        (<op> ^direction <dir)
        (<op> ^direction dir>)
In this example, it's likely that 'dir' is a meant to be a variable,
not a constant.

Visual Soar now detects issues like this and reports them to the
user during syntax checks.