Paste/Link in datamap editor menus are grayed out when clipboard is unusable
The code to check for “is this clipboard paste-able” checks for:
- is the clipboard empty or not
- if not empty, are the contents of a usable "flavor"

It can be fooled but I think it will behave properly in most cases.  Making it smarter creates a danger of delayed display of the menu or a crash when the window is opened.  I struck what I felt was the right balance between robustness and usability.

While implementing this I noticed that ctrl-c does not map to the DataMap's copyAction as it should.  It's apparently still using the system clipboard instead of the datamap's keyboard.