Fixed Bug:  Visual Soar not honoring pipe-delimeted strings as enumeration values
Example: an enumeration in the datamap whose values are:  |value one| and |value two|

VS should read this as "value one" and "value two".  Instead it was reading "|value", "one|", "|value", and "two|".  This resulted in a .dm file that couldn't be loaded.  The user was faced with a popup stating "Data Incorrectly Formatted."