Browse button robust to missing starting folder
When you create a new Soar Project and then hit the "Browse..." button, a dialog should appear.  It's not appearing for some users and this may be because the working directory no longer exists.  I've added code to notice this and use the current working dir in that case instead.