Fixed Bug:  Syntax highlighting issues and crashes on Windows
When Visual Soar is compiled with Linux and run on Windows we were seeing aberrent behavior: failed syntax highlighting and crashes.  Below is a detailed explanation of why this was happening.

The syntax highlighting code uses constants generated by the parser to figure out what colors to use.  These are constants associated with certain tokens (e.g., RARROW, SP, SYMBOLIC_CONST, etc.)  The user's preferred colors for each constant are stored in an array for fast lookup.

However, each time I modify the parser code, the values of these constants change.  (Recently I added support for some new preference operators.)  This means that syntax colors that were saved in Prefs (e.g., the registry in Windows) with a previous version of vs were now at the wrong positions.

Worse, the array might be too small.  This was because the array size was set based upon which constant was the largest.  However, if the constant values change, the new array might be smaller than the old one.  This caused syntax highlighting to crash because of array out of bounds errors.

To fix all of this, I've chosen a new, fixed size for the color syntax array that is larger than any that might be stored on any existing computer.  This wastes a trivial amount of RAM but provides a flexible and comfortable solution.

I've also added code to catch future crashes (should they occur) and just use a default color (black) for syntax highlighting.