Yet another Browse button fix (and more!)
If you are on a network drive on Windows then you have only one drive.
This was fooling VS into thinking it was on a Unix system and causing
the Browse... button on the New Project dialog to fail because it
was trying to set the current root when there were no roots listed
because a single root is treated as unix.

This, in turn, revealed other bugs with New Project which was not
writing the .vsa and .dm file on project creation

This in turn prevented the new project from being added to the
Recent list.

This also created problems with the .cfg file which didn't
exist and was causing a crash.

Overall, a ridiculous chain of bugs now fixed.  I hope no new
secret bugs in the chain but I'd not be surprised if they turned up.