WIP Confirm with user before exiting the application
Alt-X saves and closes everything in VisualSoar with no confirmation! That's
very dangerous. This introduces a confirmation dialog box before closing.

I can't merge it because there's a remaining issue: alt-x inputs ≈ on Mac, and
that input is being put into the editor before the save/close dialog opens!
I tried a couple things from the internet, but did not figure out the issue.