Improve CommentOutAction behavior
If the cursor sits at the beginning of a new line, don't comment out that line.
Don't expand the selection to the end of that line. This is standard in other
editors, and makes it easier to select some lines with "shift+up/down" and then
comment out the parts that you see visually selected. It also has a technical
logic to it, since the \n and \r belong to the same line as the preceding text,
not the following text.