Fix line commenting for selection start mid-line
If the selection starts in the middle of the line, my previous change
(cfd728acbc1aba1fae84eb71bd4cad18920d26d1) would cause the comment character to
go in the middle of the line.

The proper solution is to expand the first selected line back to its beginning,
but not to expand the last selected line forward to its end.