Refactorings towards testing (Un)CommentOutAction
Move most RuleEditor actions into dedicated files. This is just to untangle
state a bit and make the actions more testable. Add a test directory with
placeholders for the testing UncommentOutAction and CommentOutAction. The
actions that need more work to refactor were not moved. The refactorings have
been done using IntelliJ's built-in commands, so I have high confidence in the correctness.