combine edits in `(Un)CommentOutAction` for undo
Add the capability in `CustomUndoManager` to explicitly mark a series of edits
as being part of a single compound edit, based on the same `isSignificant` flag
that gets set in `CustomUndoableEdit`. Use the `AutoCloseable` interface to
prevent any future clients from forgetting to turn off the compound mode. Turn
off compound mode immediately on undo/redo so that extra edits don't get
combined.

Pass the undo manager to `UncommentOutAction` and `CommentOutAction` and
activate the compound mode while it's edits are performed.

I wish that we could also restore the selection locations after undo/redo,
but selection changes are not directly modeled as edit events that are passed
to an UndoManager in Swing, so this is the best we can do for now.