migrate away from `isSignificant` for UndoManager
Build on the example `CompoundUndoManager` provided here 25 (!) years ago:
https://tips4java.wordpress.com/2008/10/27/compound-undo-manager/

We expand the heuristics used for determining if an edit should be merged or
not, factored and simplified as much as possible. This is conceptually simpler
mechanism, doesn't require parsing UndoManager's `toString()`, and is clearly
tested by others. There's also more logic in there for fixing the selection
after undo's, though I didn't really test it so I don't know what the difference
is.