Skip (and warn for) blank lines in comments.dm
The THOR-Soar project was failing to load because the comments.dm has blank
lines, and those caused exceptions during loading (because `line.charAt(0)` was
always out of bounds). After removing the blank lines, the comments appeared
to be on the correct nodes, so we'll just ignore them. I do not know how the
blank lines showed up in the file, but I find it very worrisome.