Report errors while loading datamap
When we had an exception reading the THOR-Soar project file, the project
silently failed to load. Surround the method in a `try/catch` and report the
exception message to the user. This still isn't ideal, because we are setting
the status bar error but it immediately gets overridden by another regular
message. Fixing this would require more work; at least we now have *some*
indication that there was an error while loading, and the rest of the project
can load; I assume that's okay, given the method always returned a success flag.
A part of me is afraid that actually V-S will mess up the project somehow if we
continue without the datamap being properly loaded.