atomic feedback list management
Move the public methods for updating the status bar and feedback list into
`FeedbackManager`, and expose that in `MainFrame`. Move all of the feedback list
classes into a new package.

Support an atomic mode for the feedback list, meaning that for the duration of
the mode items are appended instead of replacing the entire list contents. Use
this mode when opening projects. There are probably other places to use it, as
well, but this addresses the current issue of project errors being hidden by
later feedback messages.