Add tests for parsing DM vertices
* Tests are parameterized so that we test both the old and new parser code
* Fix infinite while-loop issue in the old parser's `getWord`; `Reader#ready`
indicates that `read` would be non-blocking, not that read would return new
data. To find the end of the reader's data we have to watch for `-1` returned by
`read`.
* Get rid of deprecated `new Float(num)` call
* Add getters for high/low in the range vertex classes for testing. We can
test the string representation if we write tests for the individual classes,
but for testing parsing checking the field values is more appropriate.

We have to disable one test currently because the new parsing code fails it.