Fix enumeration and foreign soar vertex reading
Re-write the `readVertexSafe` method to use the `ReaderUtils` methods, since
they seem more robust than string splitting. This brings the parsing of
ENUMERATION and FOREIGN back in line with the older `readVertex` method. I still
see several issues, though:

* The line was used for testing ENUMERATION parsing is clearly malformed; each
enum val should start and end with a "|", but the values appear only to be
joined and then given one extra at the head and tail of the list. Did that line
get written by V-S? If so, then how?
* We can't handle escaped vertical bars
* I'm certain there are more potential exceptions that we are not handling. I've
marked one with a TODO. The parsing really needs one character of lookahead to
handle `\r` properly.