improve error visibility
* `restoreId` was throwing a `ClassCastException`, which is not one that we were
handling in the try/catch. Instead of trying to catch all possible errors and
ignoring them, just add them to the feedback list.
* Add an explicit note to the user that continuing with the project may cause
data loss and attention is needed. It would probably be better to have a popup
dialog because of how serious this is, but I don't want to stick a dialog deep
in the stack here. I will probably revisit this and place a dialog elsewhere,
after passing up a signal that the project appears corrupted.
* Change many error feedback messages to be shown with the error styling