WIP Exhaust all possible Soar library locations
Add a new Soar home preference. When creating a Soar kernel, catch DLL load
errors and fall back to SOAR_HOME, ../bin (in case we are inside the Soar
distribution), and finally let the user locate the Soar home (and save it in the
preference).

See https://github.com/SoarGroup/Soar/issues/491.