Add util for atomic file writing
Given a particular OS or drive combination, atomic rename may not be available,
but we do the overall write as atomically as possible. The method takes a
function for writing the file contents, rather than a string, so that we can
support streaming to the file, which is more efficient.