Use subclasses for JSON DM vertices
I played around with different configurations, but decided in the end to keep
`type` as a regular field, instead of leaving it out and only using it as
metadata for subclass resolution, for 2 reasons:

1) Having the same string in the Java class that was used for deserialization
means that we can mention these exact strings in error messages, which will be
important for usability, and

2) Using `@JsonTypeInfo(include=JsonTypeInfo.As.PROPERTY)` cause the type field
to always be printed first, despite configuring alphabetical output, and though
this may be convenient to look at, it's likely to be annoying to replicate in
other implementations, which would end up causing line noise when switching
between editors.

I do wish that Jackson supported using the enum directly in
`@JsonSubTypes.Type(name=...)`, but that turns out to actually be a Java
limitation: https://github.com/FasterXML/jackson-databind/issues/2739.