support infinity floats; remove unneeded fields
String, int and float vertices don't specify values, so we actually don't need
special classes for them at all.

Use strings to store the float range min/max so that we can parse pos/neg
infinity. Jackson actually supports using -Inf and Inf, but that's not standard
and would be a hindrance to parsing in other tools.