WIP: read project JSON
This currently throws an index out of bounds exception because the DM vertices
are sorted alphabetically but the implementation requires them to be sorted
integers (because we use Vector as a map with indices instead of keys).

So, next we have to migrate to string keys before we can make any more progress.