store serialization ID in nodes and vertices
The project JSON format does (for good reason!) not use
sequential integers that map directly to internal data structure indices as
unique IDs. It will use arbitrary strings instead.

I planned at first to just switch to using arbitrary string IDs for the internal
data structures (Graph), but, while not impossible, this was such a big change
that it would have been very difficult to test and gain confidence in its
correctness.

Instead, we just store the string ID used for project serialization next to the
existing `id` field. This isn't used when deserializing the old project formats,
so it essentially doubles the number of constructors, and in some cases you have
to be very careful of the argument ordering to ensure that you are using the
correct one.

This is kind of yucky and I wish it didn't have to be done, but I will have more
confidence in its correctness than the alternative (given a reasonable
development timebox).