normalize DM paths to forward slashes
Backslash is an escape character on Posix, and Windows understands forward
slashes, so lets normalize to forward slash so that projects work cross-
platform.

This came up because I was not able to verify or re-import a subtree for a
project developed on Windows while I was working on a Mac.