for testing only: correctly fill DM DAG from JSON
The internal data structure that hold the DM (`DirectedGraphAsAdjacencyLists`)
requires vertices to be added in order of integer ID, ascending. Failing to do
this leads to an IndexOutOfBoundsException.

For now, we sort by the parsed integer IDs, which correspond to the IDs that
were written in the VSA file format. This is just so that we can do some project
file round-tripping to make sure no information is lost. Once testing is
complete, we'll just generate the integer IDs used internally at project load
time.