working import of foreign JSON project
Implement the workflow for importing DM vertices from foreign JSON projects,
including the initial import as well as the reimport and verify actions.

This comes with a little bit of refactoring, including storing the path to the
loaded project in the `SoarWorkingMemoryModel` class, and changing some of the
load functions to be more functional (returning result objects instead of
mutating inputs).

We also fix a bug with creating new projects: if no agent path is specified,
we default to the working directory (which is determined by what folder the user
last browsed to, with "." as the default). Previously the default would be "",
which would cause the project path to be "/project-name.vsa", which would fail
to be created.