test round-trip between JSON and OperatorWindow
Fix a bug that caused edges in a foreign vertex to be mis-assigned when the
import vertex ID was the same as one of the native IDs that also had edges.

Add missing serialization of comments in layout nodes.

Fully round-trip between JSON and a loaded `OperatorWindow` (`VSTreeNode` and
`SoarWorkingMemoryModel`). Add a temporary sample file for now that uses only
number IDs. We can use the regular sample file again once we get rid of the
requirement that IDs are numbers in strings.