(de)serlialize "generated" flag
Now if many edges are generated and the user closes VisualSoar, they will still
be able to see what was generated and decide to approve it or not. This is not
supported with the VSA format.

This came up when a user had some really wonky DM entries that messed up the VSA
parser, and we assumed it was due to generating but we weren't sure.