Fix initial project backup error
First sleep and *then* try to backup. If we backup first in the while loop
before sleeping, the first attempt always fails because the operator window
hasn't actually been loaded yet.

Properly exit the loop on a thread interrupt signal.

Get rid of the `closed` signal, which was always `false`. I don't know the
proper way to handle removing the thread when the component is closed, so it
seems that we will just keep spawning these threads every time a project is
opened until we close V-S. I can't fix it right now.

Print any errors that occur during project saving.