use random IDs
instead of converting the internal integer IDs into strings. The new IDs are
UUID v4 with the dashes removed for easier editor selection.

I considered using shorter IDs but with a central store of some kind to check
for duplicates, but I think universally unique is fine and the implementation is
simple. I also considered using some deterministic IDs for the default project
elements (root, intialize-operator, name, etc.), but I think it would be
annoying to have unrelated vertices with duplicate IDs in the case of importing
from another DM, and again universal uniqueness is the simpler solution.