fix bug caused by root not being first vertex
The root always has to be assigned the internal ID of 0. If you try to add
element 1 before element 0 to a vector, it throws an array out of bounds
exception. Therefore, the internal data format requires that we add the root
first.

Update sample.json to put the root later in the vertex list and ensure this
bug is gone.