Working CLI command to check all prods against DM
Do just the minimal amount of work to get a "check all productions against
datamap" CLI action working. This required a bit of refactoring so that
everything could run headless and not hit an NPE when the `MainFrame` or
`OperatorWindow` singletons are not initialized.

Introduce the commons-cli library for the CLI parameter parsing. This is the
same library used by the Soar debugger.

Add some missing `@Override` annotations and allow IDEA to do a couple of
simplifications. Try to move some code from `io` to `nio`, from `Vector` to
`List`.

Get rid of the version 5 reading code, since the writing code for it
was never added, meaning that there can't possibly be any version 5 projects
out there. This was done here because it was in the way of one of the required
refactorings.

Support both human-readable and JSON output, where the JSON is meant for IDE/
Editor consumption (think Soar authoring plugins!).

See #47. Tests are TODO.