mark new projects as JSON
Upon saving new projects, we were always warning the user about the change of
format from .vsa to .vsa.json, which makes no sense if the project has never
been serialized before.

Fix this issue, and make the new project API a bit clearer, using a named static
method instead of one of many constructors with parameters to differentiate
them.

Add a test to ensure that new projects are marked as being in JSON format
already and, while I'm at, test the attributes and structure of a new project,
as well.