don't ask user about keeping the dialog open
With the new default being to keep the find dialog open, it doesn't make sense
to have an extra option to close it after the search finishes, since that's an
extra step for the user and closing is also only one step.

Remove the checkbox for the "Keep dialog" option everywhere, and always do the
"true" path behavior.

Reflect this change in `SearchDataMapDialog`, as well, which I had missed
before.