Fix broken/missing/wrong keyboard shortcuts
* Many listed keyboard shortcuts, such as control-F for find, did no
work. Fix this by explicitly registering shortcuts, rather than
expecting them to be registered automatically after the menu was set
up.
* Shortcuts were all using control, which is unnatural on Mac. Add a
utility method to get the key stroke, using CMD on Mac and CTRL on other
platforms. Apply this everywhere so that the keyboard shortcuts make sense for
Mac.
* Fix an NPE in `findForward()`.

See #57