use named edge for FakeTreeNodes
Previously if you had a DM node `code : integer` selected, that's how the text
would be populated in the project search box. We don't want the ` : integer`
part. To get rid of it, we add special handling for `FakeTreeNode` nodes, of
which these are an instance, and just get the edge name if the edge exists.

See #57.