Apply font size preference to everywhere possible
This means additionally that we adjust the size of the menu bars and the border
title on the feedback pane.

Place the-usable logic in a new `FontUtils` class in the components package.

I played around with using logic like this to set the font size globally
instead:

```java
UIManager.getLookAndFeelDefaults().entrySet().forEach(entry -> {
    if (entry.getKey().toString().endsWith(".font")) {
        UIManager.put(entry.getKey(), newFont);
    }
});
```

However, this only works for setting a font everywhere, meaning that we can't
set the size separatley from the font name or style. It also doesn't adjust
row spacing, etc., so the best way forward seemed to be to handle the font size
change in each component.