fix main window resize logic
* Re-tile sub-windows automatically, if configured to do so, after main window
resizing. This gets rid of the requirement for the user to explicitly request a
re-tile after they resize the window.
* Fix issue causing the feedback panel size to be miscalculated and change
unexpectedly.
* When resizing the entire application, keep the feedback panel the same size,
meaning it appears to stick to the bottom of the main frame. I believe this is
a more intuitive resize behavior. The previous behavior tried to maintain a
preferred ratio of top to bottom pane, but usually when you resize the whole
window it's because you want more editor, not more feedback pane. Generally you
want a certain number of lines shown in the feedback pane, no matter the size
of the editors. I'll leave the ratio-based preference in place, as I think it
works fine.