Fix programmatic text insertion font size
`StyleContext.getDefaultStyleContext().getStyle(StyleContext.DEFAULT_STYLE)`
gets the style that is default for all documents where it is not set. What we
rather need is the default style for the current document, which is always set
(somewhere). This fixes comment/uncomment and template insertion always
happening with the font size that the document was opened with.